added Location model and forms, made some changes to accounts directory
