+++ apps/accounts/urls.py
@@ -9,4 +9,5 @@ urlpatterns = [
     path('edit-profile/', views.edit_profile, name='edit_profile'),
     path('users/', views.view_all_users, name='view_all_users'),
     path('users/<username>/', views.view_profile, name='view_profile'),
+    path('add-location/', views.add_location, name='add_location'),
 ]
